+++ Editor/ModuleCreator.cs
@@ -1,76 +1,89 @@
 using UnityEditor;
 using UnityEngine;
+using System;
+using System.Linq;
 using System.Collections.Generic;
 using VRC.SDK3.Dynamics.PhysBone.Components;
-using System;
 
-public class ModuleCreator : Editor
+public class ModuleCreatorSettings
 {
-    private const int PRIORITY = 49;
-    
-    [MenuItem("GameObject/Module Creator/Create Module", false, PRIORITY)]
-    private static void Main(MenuCommand menuCommand)
+    public bool IncludePhysBone = true;
+    public bool IncludePhysBoneColider = true;
+
+    public bool RenameRootTransform = false;
+    public bool RemainAllPBTransforms = false;
+    public bool IncludeIgnoreTransforms = false;
+    public GameObject RootObject = null;
+}
+
+public class ModuleCreator
+{
+    private ModuleCreatorSettings Settings;
+
+    public ModuleCreator(ModuleCreatorSettings settings)
     {
-        GameObject targetObject = menuCommand.context as GameObject;
-
-        if (targetObject == null)
-        {
-            Debug.LogError("Target object is not set.");
-            return;
-        }
-
-        CheckAndCopyBones(targetObject);
+        Settings = settings;
     }
-
-    private static void CheckAndCopyBones(GameObject targetObject)
+    
+    public void CheckAndCopyBones(GameObject sourceObject)
     {   
         try
         {
-            (GameObject root, int skin_index) = CheckObjects(targetObject);
+            (GameObject root, int skin_index) = CheckObjects(sourceObject);
 
-            (GameObject new_root, string variantPath) = CopyRootObject(root, targetObject.name);
+            (GameObject new_root, string variantPath) = CopyRootObject(root, sourceObject.name);
 
-            CleanUpHierarchy(new_root, skin_index);
-
-            RemoveComponents(new_root);
+            CleanUpHierarchy(new_root, skin_index, Settings);
 
             PrefabUtility.InstantiatePrefab(new_root);
             
-            Debug.Log(variantPath + "に保存されました");
+            Debug.Log("Saved to " + variantPath);
 
         }
         catch (Exception ex)
         {
-            Debug.LogError(ex.Message);
+            Debug.LogError(ex);
+            Debug.LogError(ex.StackTrace);
         }
     }
 
-    private static (GameObject, int) CheckObjects(GameObject targetObject)
+    private GameObject CheckRoot(GameObject targetObject)
     {
+        if (Settings.RootObject) return Settings.RootObject;
         //親オブジェクトが存在するか確認
         Transform parent = targetObject.transform.parent;
         if (parent == null)
         {
-            throw new InvalidOperationException("アバター(衣装)直下のSkinnedMeshRendererがついたオブジェクトを選択してください");
+            throw new InvalidOperationException("Select the object with SkinnedMeshRenderer directly under the avatar/costume");
         }
         GameObject root = parent.gameObject;
+        return root;
+    }
 
-        //armatureがあるか確認
-        GameObject armature = null;
+    private void CheckHips(GameObject root)
+    {
+        // Check if Hips exists under the root node
+        GameObject hips = null;
         foreach (Transform child in root.transform)
         {
-            if (child.name.ToLower().StartsWith("armature"))
+            foreach (Transform grandChild in child.transform)
             {
-                armature = child.gameObject;
-                break;
+                if (grandChild.name.ToLower().StartsWith("hip"))
+                {
+                    hips = grandChild.gameObject;
+                    break;
+                }
             }
         }
-        if (armature == null)
+        if (hips == null)
         {
-            Debug.LogWarning("Armature object not found under the root object.");
-            //throw new InvalidOperationException("Armature object not found under the root object.");
+            //throw new InvalidOperationException("Hips not found under the root object.");
+            Debug.LogWarning("Hips could not be found. Merge Armature/Setup Outfit may not work properly.");
         }
+    }
 
+    private void CheckSkin(GameObject targetObject)
+    {
         //SkinnedMeshRendererがついたオブジェクトか確認
         SkinnedMeshRenderer skinnedMeshRenderer = targetObject.GetComponent<SkinnedMeshRenderer>();
         if (skinnedMeshRenderer == null)
@@ -78,6 +91,23 @@ public class ModuleCreator : Editor
             
             throw new InvalidOperationException($"{targetObject.name} does not have a SkinnedMeshRenderer.");
         }
+    }
+
+    private void Checktarget(GameObject targetObject)
+    {
+        if (targetObject == null)
+        {
+            throw new InvalidOperationException("Target object is not set.");
+        }
+
+    }
+
+    private (GameObject, int) CheckObjects(GameObject targetObject)
+    {
+        Checktarget(targetObject);
+        GameObject root = CheckRoot(targetObject);
+        CheckHips(root);
+        CheckSkin(targetObject);
 
         //skin_index: 複製先でSkinnedMeshRendererがついたオブジェクトを追跡するためのインデックス
         Transform[] AllChildren = GetAllChildren(root);
@@ -86,31 +116,62 @@ public class ModuleCreator : Editor
         return (root, skin_index);
     }
 
-    private static HashSet<GameObject> CheckBoneWeight(GameObject targetObject)
+    private HashSet<GameObject> CheckBoneWeight(GameObject targetObject)
     {   
         SkinnedMeshRenderer skinnedMeshRenderer = targetObject.GetComponent<SkinnedMeshRenderer>();
         // 指定のメッシュにウェイトを付けてるボーンの一覧を取得
         HashSet<GameObject> weightedBones = GetWeightedBones(skinnedMeshRenderer);
 
-        Debug.Log($"Bones influencing {targetObject.name}: {weightedBones.Count}/{skinnedMeshRenderer.bones.Length}");
+        Debug.Log($"Bones weighting {targetObject.name}: {weightedBones.Count}/{skinnedMeshRenderer.bones.Length}");
         return weightedBones;
     }
 
-    private static HashSet<GameObject> GetWeightedBones(SkinnedMeshRenderer skinnedMeshRenderer)
-    {   
+    private HashSet<GameObject> GetWeightedBones(SkinnedMeshRenderer skinnedMeshRenderer)
+    {
         BoneWeight[] boneWeights = skinnedMeshRenderer.sharedMesh.boneWeights;
         HashSet<GameObject> weightedBones = new HashSet<GameObject>();
+        bool hasNullBone = false;
+
         foreach (BoneWeight boneWeight in boneWeights)
         {
-            if (boneWeight.weight0 > 0) weightedBones.Add(skinnedMeshRenderer.bones[boneWeight.boneIndex0].gameObject);
-            if (boneWeight.weight1 > 0) weightedBones.Add(skinnedMeshRenderer.bones[boneWeight.boneIndex1].gameObject);
-            if (boneWeight.weight2 > 0) weightedBones.Add(skinnedMeshRenderer.bones[boneWeight.boneIndex2].gameObject);
-            if (boneWeight.weight3 > 0) weightedBones.Add(skinnedMeshRenderer.bones[boneWeight.boneIndex3].gameObject);
+            if (boneWeight.weight0 > 0)
+            {
+                Transform boneTransform = skinnedMeshRenderer.bones[boneWeight.boneIndex0];
+                if (boneTransform == null) hasNullBone = true;
+                else weightedBones.Add(boneTransform.gameObject);
+            }
+
+            if (boneWeight.weight1 > 0)
+            {
+                Transform boneTransform = skinnedMeshRenderer.bones[boneWeight.boneIndex1];
+                if (boneTransform == null) hasNullBone = true;
+                else weightedBones.Add(boneTransform.gameObject);
+            }
+
+            if (boneWeight.weight2 > 0)
+            {
+                Transform boneTransform = skinnedMeshRenderer.bones[boneWeight.boneIndex2];
+                if (boneTransform == null) hasNullBone = true;
+                else weightedBones.Add(boneTransform.gameObject);
+            }
+
+            if (boneWeight.weight3 > 0)
+            {
+                Transform boneTransform = skinnedMeshRenderer.bones[boneWeight.boneIndex3];
+                if (boneTransform == null) hasNullBone = true;
+                else weightedBones.Add(boneTransform.gameObject);
+            }
         }
+
+        if (hasNullBone)
+        {
+            throw new InvalidOperationException("Some bones weighting mesh could not be found");
+        }
+
         return weightedBones;
     }
 
-    private static (GameObject, string) CopyRootObject(GameObject root_object, string target_name)
+    private (GameObject, string) CopyRootObject(GameObject root_object, string source_name)
     {
         string base_path = $"Assets/ModuleCreator";
         if (!AssetDatabase.IsValidFolder(base_path))
@@ -126,7 +187,7 @@ public class ModuleCreator : Editor
             AssetDatabase.Refresh();
         }
 
-        string fileName = $"{target_name}_MA";
+        string fileName = $"{source_name}_MA";
         string fileExtension = "prefab";
         
         string variantPath = AssetDatabase.GenerateUniqueAssetPath(folderPath + "/" + fileName + "." + fileExtension);
@@ -134,44 +195,54 @@ public class ModuleCreator : Editor
         GameObject new_root = PrefabUtility.SaveAsPrefabAsset(root_object, variantPath);        
         if (new_root == null)
         {
-            throw new InvalidOperationException("Prefabの作成に失敗しました。");
+            throw new InvalidOperationException("Prefab creation failed.");
         }
         return (new_root, variantPath);
     }
 
-    private static void CleanUpHierarchy(GameObject new_root, int skin_index)
+    private void CleanUpHierarchy(GameObject new_root, int skin_index, ModuleCreatorSettings settings)
     {   
+        HashSet<GameObject> objectsToSave = new HashSet<GameObject>();
+
         //複製先のSkinnedMeshRendererがついたオブジェクトを取得
         Transform[] AllChildren = GetAllChildren(new_root);
         GameObject skin = AllChildren[skin_index].gameObject;
-        skin.SetActive(true);
-        skin.tag = "Untagged"; 
+        objectsToSave.Add(skin);
 
         HashSet<GameObject> weightedBones = CheckBoneWeight(skin);
-        HashSet<Transform> All_PB_Transforms = Find_PB_Transforms(new_root, weightedBones);
-        CheckAndDeleteRecursive(new_root, weightedBones, skin, All_PB_Transforms);
+        objectsToSave.UnionWith(weightedBones);
+
+        if (settings.IncludePhysBone == true) 
+        {
+            HashSet<GameObject> PhysBoneObjects = FindPhysBoneObjects(new_root, weightedBones);
+            objectsToSave.UnionWith(PhysBoneObjects);
+        }
+
+        CheckAndDeleteRecursive(new_root, objectsToSave);
     }
 
-    private static void CheckAndDeleteRecursive(GameObject obj, HashSet<GameObject> weightedBones, GameObject skin, HashSet<Transform> All_PB_Transforms)
+    private void CheckAndDeleteRecursive(GameObject obj, HashSet<GameObject> objectsToSave)
     {   
         List<GameObject> children = GetChildren(obj);
 
         // 子オブジェクトに対して再帰的に処理を適用
         foreach (GameObject child in children)
         {   
-            CheckAndDeleteRecursive(child, weightedBones, skin, All_PB_Transforms);
+            CheckAndDeleteRecursive(child, objectsToSave);
         }
 
         // 削除しない条件
-        if (obj == skin || All_PB_Transforms.Contains(obj.transform) || weightedBones.Contains(obj) || obj.transform.childCount != 0)
+        if (objectsToSave.Contains(obj) || obj.transform.childCount != 0)
         {
-            //RemoveComponents(obj);
+            ActivateObject(obj);
+            RemoveComponents(obj);
             return;
         }
-        DestroyImmediate(obj, true);
+        
+        UnityEngine.Object.DestroyImmediate(obj, true);
     }
 
-    private static List<GameObject> GetChildren(GameObject parent)
+    private List<GameObject> GetChildren(GameObject parent)
     {
         List<GameObject> children = new List<GameObject>();
         foreach (Transform child in parent.transform)
@@ -187,99 +258,141 @@ public class ModuleCreator : Editor
         return children;
     }
 
-    private static void RemoveComponents(GameObject targetGameObject)
+    private void ActivateObject(GameObject obj)
     {
-        Component[] components = targetGameObject.GetComponents<Component>();
+        obj.SetActive(true);
+        obj.tag = "Untagged"; 
+    }
 
-        foreach (Component component in components)
+    private void RemoveComponents(GameObject targetGameObject)
+    {
+        // Componentを列挙し、Transform、SkinnedMeshRenderer、(VRCPhysBone、VRCPhysBoneCollider)以外を削除
+        List<Component> componentsToRemove;
+        if (Settings.IncludePhysBone == true)
         {
-            // コンポーネントがTransform以外の場合、削除
-            if (!(component is Transform))
-            {
-                DestroyImmediate(component, true);
-            }
+            componentsToRemove = targetGameObject.GetComponents<Component>()
+                .Where(c => !(c is Transform) && !(c is SkinnedMeshRenderer)&& !(c is VRCPhysBone) && !(c is VRCPhysBoneCollider))
+                .ToList();
+        }
+        else
+        {
+            componentsToRemove = targetGameObject.GetComponents<Component>()
+                .Where(c => !(c is Transform) && !(c is SkinnedMeshRenderer))
+                .ToList();
+        }
+
+        foreach (var component in componentsToRemove)
+        {
+            UnityEngine.Object.DestroyImmediate(component, true);
         }
     }
 
-    //配下のオブジェクトにおいて、枝分かれしない限り再帰的に追加する
-    //単一のPBコンポーネントの影響下に複数のメッシュが紐づけられている際に不要なPB Transformsが適切に削除されるようになる
-    //commit: ddc51eaf8b9cacdac632962daa2bfe3e2ba51c4c
-    public static void AddSingleChildRecursive(Transform transform, HashSet<Transform> result)
-    {
-        result.Add(transform);
+    private void AddSingleChildRecursive(Transform transform, HashSet<GameObject> result, HashSet<Transform> ignoreTransforms)
+    {   
+        if (Settings.IncludeIgnoreTransforms == false && ignoreTransforms.Contains(transform)) return;
+        result.Add(transform.gameObject);   
         if (transform.childCount == 1)
         {
             Transform child = transform.GetChild(0);
-            AddSingleChildRecursive(child, result);
+            AddSingleChildRecursive(child, result, ignoreTransforms);
         }
     }
 
-    //VRCPhysBoneとVRCPhysBoneColliderを検索し、削除対象から除外するためのHashSetを返す
-    private static HashSet<Transform> Find_PB_Transforms(GameObject root, HashSet<GameObject> weightedBones)
-    {   
-        HashSet<Transform> All_PB_Transforms = new HashSet<Transform>();
+    private HashSet<Transform> GetIgnoreTransforms(VRCPhysBone physBone)
+    {
+        HashSet<Transform> AffectedIgnoreTransforms = new HashSet<Transform>();
 
-        VRCPhysBone[] physBones = root.GetComponentsInChildren<VRCPhysBone>(true);
-        foreach (VRCPhysBone physBone in physBones)
+        foreach (Transform ignoreTransform in physBone.ignoreTransforms)
         {   
-            if (physBone.rootTransform == null) 
-            {   
-                physBone.rootTransform = physBone.transform;
-            }
-            
-            Transform[] PB_Transforms = GetAllChildren(physBone.rootTransform.gameObject);
-
-            HashSet<Transform> weighted_PB_Transforms = new HashSet<Transform>();
-
-            //保持されるべき一部の親オブジェクトは現状追加されていないが、削除時にchildCount == 0の条件が含まれるので必要な親オブジェクトは実際には保持される
-            foreach (Transform PB_Transform in PB_Transforms)
-            {
-                if (weightedBones.Contains(PB_Transform.gameObject))
-                {   
-                    HashSet<Transform> result = new HashSet<Transform>();
-                    AddSingleChildRecursive(PB_Transform, result);
-                    weighted_PB_Transforms.UnionWith(result);
-                }
-            }
-
-            //有効なPBだった場合
-            if (weighted_PB_Transforms.Count > 0)
-            {
-                //MAの仕様に反し衣装側のPBを強制
-                //physBone.rootTransform.name = $"{physBone.rootTransform.name}.1";
-
-                All_PB_Transforms.Add(physBone.transform);
-                All_PB_Transforms.UnionWith(weighted_PB_Transforms);
-
-                foreach (VRCPhysBoneCollider collider in physBone.colliders)
-                {   
-                    if (collider.rootTransform == null) 
-                    {   
-                        collider.rootTransform = collider.transform;
-                    }
-
-                    All_PB_Transforms.Add(collider.transform);
-                    All_PB_Transforms.Add(collider.rootTransform);
-                }
-            }            
+            if (ignoreTransform == null) continue;
+            Transform[] AffectedIgnoreTransform = GetAllChildren(ignoreTransform.gameObject);
+            AffectedIgnoreTransforms.UnionWith(AffectedIgnoreTransform);
         }
 
-        //不要なコライダーを削除
-        VRCPhysBoneCollider[] colliders = root.GetComponentsInChildren<VRCPhysBoneCollider>(true);
-        foreach (VRCPhysBoneCollider collider in colliders)
+        return AffectedIgnoreTransforms;
+    }
+
+    private HashSet<GameObject> FindPhysBoneObjects(GameObject root, HashSet<GameObject> weightedBones)
+    {
+        var physBoneObjects = new HashSet<GameObject>();
+
+        foreach (VRCPhysBone physBone in root.GetComponentsInChildren<VRCPhysBone>(true))
         {
-            if (All_PB_Transforms.Contains(collider.transform))
+            if (physBone.rootTransform == null) physBone.rootTransform = physBone.transform;
+            var weightedPBObjects = GetWeightedPhysBoneObjects(physBone, weightedBones);
+            if (weightedPBObjects.Count > 0)
+            {
+                //MAの仕様に反し衣装側のPBを強制
+                if (Settings.RenameRootTransform == true)
+                {
+                    physBone.rootTransform.name = $"{physBone.rootTransform.name}.1";
+                }
+
+                physBoneObjects.Add(physBone.gameObject);
+                physBoneObjects.UnionWith(weightedPBObjects);
+
+                if (Settings.IncludePhysBoneColider == true)
+                {
+                    foreach (VRCPhysBoneCollider collider in physBone.colliders)
+                    {
+                        if (collider == null) continue;
+                        if (collider.rootTransform == null) collider.rootTransform = collider.transform;
+                        physBoneObjects.Add(collider.gameObject);
+                        physBoneObjects.Add(collider.rootTransform.gameObject);
+                    }
+                }
+            }
+            else UnityEngine.Object.DestroyImmediate(physBone, true);
+        }
+
+        if (Settings.IncludePhysBoneColider == true) RemoveUnusedPhysBoneColliders(root, physBoneObjects);
+        return physBoneObjects;
+    }
+
+    private HashSet<GameObject> GetWeightedPhysBoneObjects(VRCPhysBone physBone, HashSet<GameObject> weightedBones)
+    {
+        var WeightedPhysBoneObjects = new HashSet<GameObject>();
+        HashSet<Transform> ignoreTransforms = GetIgnoreTransforms(physBone);
+
+        Transform[] allchildren = GetAllChildren(physBone.rootTransform.gameObject);
+
+        foreach (Transform child in allchildren)
+        {
+            if (weightedBones.Contains(child.gameObject))
+            {
+                if (Settings.RemainAllPBTransforms == true)
+                {
+                    WeightedPhysBoneObjects.UnionWith(allchildren.Select(t => t.gameObject));
+                    break;
+
+                }
+                HashSet<GameObject> result = new HashSet<GameObject>();
+                AddSingleChildRecursive(child, result, ignoreTransforms);
+                WeightedPhysBoneObjects.UnionWith(result);
+            }
+        }
+
+        return WeightedPhysBoneObjects;
+    }
+
+    private void RemoveUnusedPhysBoneColliders(GameObject root, HashSet<GameObject> physBoneObjects)
+    {
+        foreach (VRCPhysBoneCollider collider in root.GetComponentsInChildren<VRCPhysBoneCollider>(true))
+        {
+            if (physBoneObjects.Contains(collider.gameObject))
             {
                 //MAの仕様に反し衣装側のPBCを強制
-               //collider.rootTransform.name = $"{collider.rootTransform.name}.1";
+                if (Settings.RenameRootTransform == true)
+                {
+                    collider.rootTransform.name = $"{collider.rootTransform.name}.1";
+                    //Debug.Log(collider.rootTransform.name);
+                }
             }
             else
             {
-                DestroyImmediate(collider, true);
+                UnityEngine.Object.DestroyImmediate(collider, true);
             }
         }
-
-        return All_PB_Transforms;
     }
 
 }